windows10 color

+++ engine/VertexArray.cpp
@@ -43,7 +43,7 @@ void VertexArray::addBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
     bind();
     vb.bind();
     const std::vector<VertexBufferElement>& elements = layout.getElements();
-    unsigned long int offset = 0;
+    unsigned long long int offset = 0;
     for (unsigned int i = 0; i < elements.size(); i++){
         const VertexBufferElement& element = elements[i];
         glEnableVertexAttribArray(i);
